Implement the beginnings of a plugin scheme.
The program now looks for any importable modules/packages whose
names begin with "filescan_". For each new or updated file, that
module's process function is called with two arguments. The first
is the connection object being used to implemenbt the storage
interface and the second is the (updated or new) Location object
describing the file in question.